@@ -11,104 +11,109 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "accesscontrol": "^1.0.0",
-    "autosize": "^3.0.17",
+    "accesscontrol": "^1.0.1",
+    "autosize": "^3.0.19",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "~1.15.2",
     "connect-redis": "^3.1.0",
     "cookie-parser": "~1.4.3",
-    "debug": "~2.2.0",
+    "debug": "~2.3.3",
     "emojify.js": "^1.1.0",
     "emojione": "^2.2.6",
     "emojione-picker": "^1.1.0",
     "es6-promise": "^4.0.5",
     "express": "^4.14.0",
-    "express-session": "^1.14.1",
+    "express-session": "^1.14.2",
     "fetch": "^1.1.0",
-    "helmet": "^2.3.0",
+    "fine-uploader": "^5.11.9",
+    "gm": "^1.23.0",
+    "helmet": "^3.1.0",
+    "history": "^4.4.0",
     "identicon.js": "^1.3.0",
     "isomorphic-fetch": "^2.2.1",
     "jdenticon": "^1.3.2",
     "js-emoji-fork": "^1.0.0",
     "jsonwebtoken": "^7.1.9",
-    "lodash": "^4.16.4",
-    "moment": "^2.15.1",
+    "lodash": "^4.17.2",
+    "moment": "^2.17.0",
     "morgan": "~1.7.0",
-    "mysql": "^2.11.1",
+    "multer": "^1.2.0",
+    "mysql": "^2.12.0",
     "node-emoji": "^1.4.1",
     "nodemailer": "^2.6.4",
     "normalize.css": "^5.0.0",
     "normalizr": "^2.2.1",
     "passport": "^0.3.2",
-    "passport-jwt": "^2.1.0",
+    "passport-jwt": "^2.2.1",
     "passport-local": "^1.0.0",
     "pg": "^6.1.0",
     "pg-hstore": "^2.3.2",
-    "react": "^15.3.2",
-    "react-addons-css-transition-group": "^15.3.2",
-    "react-addons-shallow-compare": "^15.3.2",
+    "rc-upload": "^2.2.0",
+    "react": "^15.4.1",
+    "react-addons-css-transition-group": "^15.4.1",
+    "react-addons-shallow-compare": "^15.4.1",
     "react-autobind": "^1.0.6",
     "react-custom-scrollbars": "^4.0.0",
     "react-dnd": "^2.1.4",
     "react-dnd-html5-backend": "^2.1.2",
-    "react-dom": "^15.3.2",
+    "react-dom": "^15.4.1",
     "react-emoji-picker": "^1.0.12",
     "react-emojione": "^2.0.0",
-    "react-helmet": "^3.1.0",
+    "react-helmet": "^3.2.2",
     "react-hot-loader": "^1.3.0",
     "react-input-autosize": "^1.1.0",
     "react-modal": "^1.5.2",
     "react-redux": "^5.0.0-beta.1",
-    "react-router": "^3.0.0-alpha.3",
+    "react-router": "^4.0.0-alpha.6",
+    "react-router-addons-controlled": "0.0.1",
     "react-select": "^1.0.0-rc.2",
-    "react-sortable-hoc": "^0.1.0",
+    "react-sortable-hoc": "^0.2.0",
     "react-tether": "^0.5.2",
-    "react-twemoji": "0.0.4",
+    "react-twemoji": "0.0.5",
     "redux": "^3.6.0",
     "redux-thunk": "*",
     "remarkable": "^1.7.1",
     "reselect": "^2.5.4",
-    "sequelize": "^3.24.3",
-    "serve-favicon": "~2.3.0",
+    "sequelize": "^3.27.0",
+    "serve-favicon": "~2.3.2",
     "shallowequal": "^0.2.2",
-    "socket.io": "^1.5.0",
-    "socket.io-client": "^1.5.0",
+    "socket.io": "^1.6.0",
     "socketio-jwt": "^4.5.0",
-    "validator": "^6.0.0",
-    "velocity-animate": "^1.2.3",
+    "validator": "^6.1.0",
+    "velocity-animate": "^1.3.1",
     "wdt-emoji-bundle": "^0.1.3",
     "ws": "^1.1.1"
   },
   "devDependencies": {
-    "babel-cli": "^6.16.0",
-    "babel-core": "^6.17.0",
-    "babel-eslint": "^7.0.0",
-    "babel-loader": "^6.2.5",
+    "babel-cli": "^6.18.0",
+    "babel-core": "^6.18.2",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^6.2.8",
     "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.16.0",
+    "babel-preset-es2015": "^6.18.0",
     "babel-preset-nodejs6": "^1.0.0",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-0": "^6.16.0",
     "chai": "^3.5.0",
     "colors": "^1.1.2",
-    "compression-webpack-plugin": "^0.3.1",
-    "css-loader": "^0.25.0",
-    "enzyme": "^2.5.1",
+    "compression-webpack-plugin": "^0.3.2",
+    "css-loader": "^0.26.0",
+    "enzyme": "^2.6.0",
     "expect": "^1.20.2",
     "file-loader": "^0.9.0",
     "jsdom": "^9.8.3",
     "json-loader": "^0.5.4",
     "mocha": "^3.1.2",
-    "node-sass": "^3.10.1",
-    "react-addons-test-utils": "^15.3.2",
-    "react-hot-loader": "^3.0.0-beta.5",
+    "node-sass": "^3.13.0",
+    "react-addons-test-utils": "^15.4.1",
+    "react-hot-loader": "^3.0.0-beta.6",
     "redux-devtools": "^3.3.1",
-    "redux-logger": "^2.7.0",
+    "redux-logger": "^2.7.4",
     "sass-loader": "^4.0.2",
     "sinon": "^1.17.6",
     "style-loader": "^0.13.1",
-    "webpack": "^1.13.2",
+    "webpack": "^1.13.3",
     "webpack-dev-middleware": "^1.8.4",
-    "webpack-hot-middleware": "^2.13.0"
+    "webpack-hot-middleware": "^2.13.2"
   }
 }
