@@ -3,34 +3,43 @@ import { connect } from 'react-redux';
 import * as Actions from '../actions/index';
 import MessageList from '../components/MessageList';
 import autoBind from 'react-autobind';
-import { withRouter } from 'react-router';
 import { socket } from '../actions/common/socketEvents';
 import LoadMoreBlock from '../components/LoadMoreBlock';
 import shallowEqual from 'shallowequal';
 import slice from 'lodash/slice';
 import uniq from 'lodash/uniq';
 import findKey from 'lodash/findKey';
+import isEqual from 'lodash/isEqual';
 
 class MessageListContainer extends Component {
-  constructor(props){
-    super(props);
-    autoBind(this);
-    this.state={currentIds:[]};
-    this.registeredElements={};
+  state = { currentIds:[], curChannel:null };
+  registeredElements = {};
+  shouldComponentUpdate(nextProps, nextState){
+      return !isEqual(this.props, nextProps)
+       || !isEqual(this.state, nextState);
   }
   componentWillMount(){
-    this.props.fetchMessages(this.props.params.channelId)
+    const {channelId} = this.props;
+    this.props.fetchMessages(channelId)
     .then(()=> this.messageList.scrollView.scrollToBottom() );
-    this.props.selectChannel(this.props.params.channelId);
   }
   componentDidMount(){
     this.props.onMount(this.messageList);
   }
-  componentWillReceiveProps(nextProps){
-    const {channelId:prevId} = this.props.params;
-    const {channelId:nextId} = nextProps.params;
-
-    if (prevId !== nextId) {
+  componentDidUpdate(prevProps, prevState){
+    // const {channelId:prevId} = nextProps;
+    // const {channelId:nextId} = this.props;
+    // // console.info(prevId,nextId,prevId == nextId);
+    // if (prevId != nextId) {
+    //   this.handleSwitchChannel(prevId,nextId);
+    // }
+  }
+    componentWillReceiveProps(nextProps){
+    const {channelId:prevId} = this.props;
+    const {channelId:nextId} = nextProps;
+    console.info(prevId,nextId,this.state.curChannel,prevId == nextId);
+    if (prevId !== nextId ) {
+      // this.setState({curChannel :nextId });
       this.handleSwitchChannel(prevId,nextId);
     }
 
@@ -39,30 +48,18 @@ class MessageListContainer extends Component {
 
     const nextLastId = nextMsgIds.slice(-1)[0];
     const prevLastId = prevMsgIds.slice(-1)[0];
-
-    // if (prevMsgIds.length && nextMsgIds.length > prevMsgIds.length && nextLastId !== prevLastId){
-    //     const message = nextProps.messages[nextLastId];
-    //     console.log('new MESSAGE CATCHED!',message);
-    //
-    //     if(this.messageList.isScrollOnBottom()){
-    //         console.log('крутим вниз');
-    //         this.messageList.scrollView.scrollTop(this.messageList.scrollView.getScrollHeight()-5);
-    //     }
-    // }
-
   }
-  handleSwitchChannel(prevId,nextId){
+  handleSwitchChannel=(prevId,nextId)=>{
+    console.info('handleSwitchChannel');
     const channelInfo = {
       channelId: prevId,
       scrollPosition: this.messageList.scrollView.getScrollTop(),
       firstVisibleId: this.findFirstVisibleId()
     };
     this.props.saveScrollPosition(channelInfo);
-    // this.props.saveLastVisibleMessage(prevId);
-    this.props.selectChannel(nextId);
+
     this.props.fetchMessages(nextId).then(()=>{
 
-      this.props.unsetChannelHasNewMessages(nextId);
       const { scrollPosition, firstVisibleId } = this.props;
       if (scrollPosition && firstVisibleId){
         this.messageList.scrollView.scrollTop(60);
@@ -72,62 +69,61 @@ class MessageListContainer extends Component {
         this.messageList.scrollView.scrollToBottom();
       }
     });
+    this.props.unsetChannelHasNewMessages(nextId);
+
   }
 
-  loadMoreBefore(oldScrollHeight){
+  loadMoreBefore=(oldScrollHeight)=>{
     const view = this.messageList.scrollView;
-    this.props.loadMoreBefore(this.props.params.channelId)
+    this.props.loadMoreBefore(this.props.channelId)
     .then(()=> view.scrollTop(view.getScrollHeight() - oldScrollHeight));
   }
 
-  loadMoreAfter(){
-    this.props.loadMoreAfter(this.props.params.channelId);
+  loadMoreAfter=()=>{
+    this.props.loadMoreAfter(this.props.channelId);
   }
-  regElem(element){
+  regElem=(element)=>{
     this.registeredElements[element.id] = element;
   }
-  unregElem(element){
+  unregElem=(element)=>{
     delete this.registeredElements[element.id];
   }
-  findFirstVisibleId(){
+  findFirstVisibleId=()=>{
     return +findKey(this.registeredElements,({elem})=>
     elem.offsetTop + elem.offsetHeight > this.messageList.scrollView.getScrollTop());
     // elem.getBoundingClientRect().bottom > 0
   }
-  onMessageEdit(id){
-    console.log('onMessageEdit <<<<<<<',id);
+  onMessageEdit=(id)=>{
     this.props.setEditableMessage(id);
   }
-  onMessageDelete(id){
-    console.log('onMessageEdit <<<<<<<',id);
+  onMessageDelete=(id)=>{
     this.props.deleteMessage(id);
   }
-  onSaveMessageEdit(id,text){
-    console.log('onSaveMessageEdit <<<<<<<',id,text);
+  onSaveMessageEdit=(id,text)=>{
     if(text){
       this.props.unsetEditableMessage();
       this.props.editMessage(id,text);
     }
   }
-  onCancelMessageEdit(){
+  onCancelMessageEdit=()=>{
     this.props.unsetEditableMessage();
-    console.log('onCancelMessageEdit <<<<<<<');
   }
-  onScrollStop(){
+  onScrollStop=()=>{
   }
   render(){
     const {
       messages, allMessagesIds, messagesIsFetching, users, slice,
-      editableMessage,currentUser
+      editableMessageId,currentUser
     } = this.props;
+    console.log('MessageListContainer');
     return (
       <MessageList
         onSaveMessageEdit={this.onSaveMessageEdit}
         onCancelMessageEdit={this.onCancelMessageEdit}
         onMessageEdit={this.onMessageEdit}
         onMessageDelete={this.onMessageDelete}
-        editableMessage={editableMessage}
-        currentUser={currentUser}
+        editableMessageId={editableMessageId}
+        currentUser={currentUserId}
         onScrollStop={this.onScrollStop}
         onMountMessage={this.regElem}
         onUnmountMessage={this.unregElem}
@@ -143,7 +139,7 @@ class MessageListContainer extends Component {
   }
 }
 const mapStateToProps = (state) =>{
-  const channel  = state.pagination.idsByChannel[state.ui.selectedChannel];
+  const channel  = state.pagination.idsByChannel[state.ui.params.channelId];
   const {
     ids:messagesIds = [], scrollPosition = null , lastVisibleMessage={},
     firstVisibleId , slice = []
@@ -151,8 +147,9 @@ const mapStateToProps = (state) =>{
 
   const messages = state.entities.messages.items;
   return {
-    currentUser:state.auth.user,
-    editableMessage:state.ui.editableMessage,
+    channelId:state.ui.params.channelId,
+    currentUserId:state.auth.user,
+    editableMessageId:state.ui.editableMessage,
     slice,
     allMessagesIds: messagesIds,
     users: state.entities.users.items,
@@ -163,4 +160,4 @@ const mapStateToProps = (state) =>{
   };
 };
 
-export default withRouter(connect(mapStateToProps, Actions)(MessageListContainer));
+export default connect(mapStateToProps, Actions)(MessageListContainer);
