@@ -1,30 +1,30 @@
-import types from './common/types';
+import type from './common/types';
 import { normalize, arrayOf } from "normalizr";
-import { schemas } from "./common/schemas";
+import * as schemas from "./common/schemas";
 import API from "./common/API/API";
 import { checkStatus } from "./common/check_status_response";
 
 //-----------Login--------------------------------------
 export const requestLogin = () =>( {
-    type: types.LOGIN_REQUEST,
-    isFetching: true,
-    isAuthenticated: false,
-    logtest:true,
-    errors:"",
+  type: type.LOGIN_REQUEST,
+  isFetching: true,
+  isAuthenticated: false,
+  logtest:true,
+  errors:"",
 });
 
 export const receiveLogin = payload =>( {
-    type: types.LOGIN_SUCCESS,
-    isFetching: false,
-    isAuthenticated: true,
-    token: payload.token
+  type: type.LOGIN_SUCCESS,
+  isFetching: false,
+  isAuthenticated: true,
+  token: payload.token
 });
 
 export const loginError =  message  =>( {
-    type: types.LOGIN_FAILURE,
-    isFetching: false,
-    isAuthenticated: false,
-    message
+  type: type.LOGIN_FAILURE,
+  isFetching: false,
+  isAuthenticated: false,
+  message
 });
 
 
@@ -39,52 +39,52 @@ export const loginUser = creds => async (dispatch) =>{
 
   } catch (err) {
     err.response.json().then(body=>{
-        // console.log('loginUser', body, err.response.status);
-        dispatch(loginError(body));
+      // console.log('loginUser', body, err.response.status);
+      dispatch(loginError(body));
     });
   }
 };
 //-----------Logout--------------------------------------
 export const requestLogout = () =>( {
-    type: types.LOGOUT_REQUEST,
-    isFetching: true,
-    isAuthenticated: true
+  type: type.LOGOUT_REQUEST,
+  isFetching: true,
+  isAuthenticated: true
 });
 
 export const receiveLogout = () =>( {
-    type: types.LOGOUT_SUCCESS,
-    isFetching: false,
-    isAuthenticated: false
+  type: type.LOGOUT_SUCCESS,
+  isFetching: false,
+  isAuthenticated: false
 });
 
 export const logoutUser = () => dispatch=>{
-    dispatch(requestLogout());
-    localStorage.removeItem('token');
-    dispatch(receiveLogout());
+  dispatch(requestLogout());
+  localStorage.removeItem('token');
+  dispatch(receiveLogout());
 };
 
 
 //-----------Signup--------------------------------------
 export const requestSignup = () =>( {
-    type: types.SIGNUP_REQUEST,
-    isFetching: true,
-    isAuthenticated: false,
-    logtest:true
-    // creds
+  type: type.SIGNUP_REQUEST,
+  isFetching: true,
+  isAuthenticated: false,
+  logtest:true
+  // creds
 });
 
 export const receiveSignup = payload =>( {
-    type: types.SIGNUP_SUCCESS,
-    isFetching: false,
-    isAuthenticated: true,
-    token: payload.token
+  type: type.SIGNUP_SUCCESS,
+  isFetching: false,
+  isAuthenticated: true,
+  token: payload.token
 });
 
 export const SignupError = message =>( {
-    type: types.SIGNUP_FAILURE,
-    isFetching: false,
-    isAuthenticated: false,
-    message
+  type: type.SIGNUP_FAILURE,
+  isFetching: false,
+  isAuthenticated: false,
+  message
 });
 
 export const signupUser = creds => async (dispatch) =>{
@@ -98,8 +98,8 @@ export const signupUser = creds => async (dispatch) =>{
 
   } catch (err) {
     err.response.json().then(body=>{
-        console.error('SignupError', body, err.response.status);
-        dispatch(SignupError(body));
+      console.error('SignupError', body, err.response.status);
+      dispatch(SignupError(body));
     });
   }
 };
@@ -109,16 +109,16 @@ export const signupUser = creds => async (dispatch) =>{
 //---------------------current-user-----------------
 
 export const fetchCurrentUserRequest = () =>({
-  type:types.FETCH_CURRENT_USER_REQUEST,
+  type:type.FETCH_CURRENT_USER_REQUEST,
   isFetching:true
 });
-export const fetchCurrentUserSuccess = (user) =>({
-  type:types.FETCH_CURRENT_USER_SUCCESS,
+export const fetchCurrentUserSuccess = (payload) =>({
+  type:type.FETCH_CURRENT_USER_SUCCESS,
   isFetching:false,
-  user
+  payload
 });
 export const fetchCurrentUserFailure = (error) =>({
-  type:types.FETCH_CURRENT_USER_FAILURE,
+  type:type.FETCH_CURRENT_USER_FAILURE,
   isFetching:false,
   error
 });
@@ -127,9 +127,33 @@ export const fetchCurrentUser = () => async (dispatch) =>{
   try {
     dispatch(fetchCurrentUserRequest());
     const user = await API.User.get();
-    dispatch(fetchCurrentUserSuccess(user));
+    const payload = normalize(user, schemas.user);
+
+    console.log(payload, 'fetchCurrentUser');
+
+    dispatch(fetchCurrentUserSuccess(payload));
+
     return user;
   } catch (e) {
-    dispatch(fetchCurrentUserFailure(e))
+    dispatch(fetchCurrentUserFailure(e));
   }
-}
+};
+
+//---------------------upload avatar-----------------
+
+export const uploadUserAvatarRequest = () =>({
+  type:type.UPLOAD_USER_AVATAR_REQUEST,
+  isFetching:true,
+  error:null
+});
+export const uploadUserAvatarFailure = (error) =>({
+  type:type.UPLOAD_USER_AVATAR_FAILURE,
+  isFetching:false,
+  error
+});
+export const uploadUserAvatarSuccess = (avatar, userId) =>({
+  type:type.UPLOAD_USER_AVATAR_SUCCESS,
+  isFetching:false,
+  userId,
+  avatar
+});
