@@ -1,15 +1,15 @@
 import { Schema, arrayOf, valuesOf } from 'normalizr';
 
-export const user = new Schema('users',{ idAttribute: 'id' });
-export const server = new Schema('servers',{ idAttribute: 'id' });
-export const message = new Schema('messages',{ idAttribute: 'id' });
-export const channel = new Schema('channels',{ idAttribute: 'id' });
+export const user = new Schema('users');
+export const server = new Schema('servers');
+export const message = new Schema('messages');
+export const channel = new Schema('channels');
 
-export const action = new Schema('actions', { idAttribute: 'id' });
-export const resource = new Schema('resources', { idAttribute: 'id' });
-export const permission = new Schema('permissions', { idAttribute: 'id' });
-export const attribute = new Schema('attributes', { idAttribute: 'id' });
-export const role = new Schema('roles', { idAttribute: 'id' });
+export const action = new Schema('actions');
+export const resource = new Schema('resources');
+export const permission = new Schema('permissions');
+export const attribute = new Schema('attributes');
+export const role = new Schema('roles');
 
 
 role.define({
@@ -28,5 +28,5 @@ message.define({
 });
 server.define({
     channels:arrayOf(channel),
-    serverRoles:arrayOf(role)
+    roles:arrayOf(role)
 });
