@@ -1,5 +1,8 @@
 const types = {
+  //ui
+  TOGGLE_MODAL: 'TOGGLE_MODAL',
 
+  //auth
   LOGIN_REQUEST: 'LOGIN_REQUEST',
   LOGIN_SUCCESS: 'LOGIN_SUCCESS',
   LOGIN_FAILURE: 'LOGIN_FAILURE',
@@ -12,7 +15,6 @@ const types = {
   LOGOUT_SUCCESS: 'LOGOUT_SUCCESS',
   LOGOUT_FAILURE: 'LOGOUT_FAILURE',
 
-  UPDATE_USERS:'UPDATE_USERS',
   UPDATE_USERS_ONLINE:'UPDATE_USERS_ONLINE',
 
   RECEIVE_TYPING:'RECEIVE_TYPING',
@@ -60,6 +62,15 @@ const types = {
   START_TYPING:'START_TYPING',
   STOP_TYPING:'STOP_TYPING',
 
+  //user
+  EDIT_USER_REQUEST: 'EDIT_USER_REQUEST',
+  EDIT_USER_SUCCESS: 'EDIT_USER_SUCCESS',
+  EDIT_USER_FAILURE: 'EDIT_USER_FAILURE',
+
+  UPLOAD_USER_AVATAR_REQUEST: 'UPLOAD_USER_AVATAR_REQUEST',
+  UPLOAD_USER_AVATAR_SUCCESS: 'UPLOAD_USER_AVATAR_SUCCESS',
+  UPLOAD_USER_AVATAR_FAILURE: 'UPLOAD_USER_AVATAR_FAILURE',
+
   FETCH_CURRENT_USER_REQUEST: 'FETCH_CURRENT_USER_REQUEST',
   FETCH_CURRENT_USER_SUCCESS: 'FETCH_CURRENT_USER_SUCCESS',
   FETCH_CURRENT_USER_FAILURE: 'FETCH_CURRENT_USER_FAILURE',
@@ -79,6 +90,10 @@ const types = {
   UNSET_EDITABLE_MESSAGE: 'UNSET_EDITABLE_MESSAGE',
 
   //servers
+  CREATE_SERVER_REQUEST: 'CREATE_SERVER_REQUEST',
+  CREATE_SERVER_SUCCESS: 'CREATE_SERVER_SUCCESS',
+  CREATE_SERVER_FAILURE: 'CREATE_SERVER_FAILURE',
+
   FETCH_SERVERS_REQUEST: 'FETCH_SERVERS_REQUEST',
   FETCH_SERVERS_SUCCESS: 'FETCH_SERVERS_SUCCESS',
   FETCH_SERVERS_FAILURE: 'FETCH_SERVERS_FAILURE',
@@ -88,10 +103,38 @@ const types = {
   EDIT_SERVERS_ORDER_SUCCESS: 'EDIT_SERVERS_ORDER_SUCCESS',
   EDIT_SERVERS_ORDER_FAILURE: 'EDIT_SERVERS_ORDER_FAILURE',
 
+  UPLOAD_SERVER_ICON_REQUEST: 'UPLOAD_SERVER_ICON_REQUEST',
+  UPLOAD_SERVER_ICON_SUCCESS: 'UPLOAD_SERVER_ICON_SUCCESS',
+  UPLOAD_SERVER_ICON_FAILURE: 'UPLOAD_SERVER_ICON_FAILURE',
+
   //roles
   FETCH_ROLES_REQUEST: 'FETCH_ROLES_REQUEST',
   FETCH_ROLES_SUCCESS: 'FETCH_ROLES_SUCCESS',
-  FETCH_ROLES_FAILURE: 'FETCH_ROLES_FAILURE'
+  FETCH_ROLES_FAILURE: 'FETCH_ROLES_FAILURE',
+
+  CREATE_ROLE_REQUEST: 'CREATE_ROLE_REQUEST',
+  CREATE_ROLE_SUCCESS: 'CREATE_ROLE_SUCCESS',
+  CREATE_ROLE_FAILURE: 'CREATE_ROLE_FAILURE',
+
+  DELETE_ROLE_REQUEST: 'DELETE_ROLE_REQUEST',
+  DELETE_ROLE_SUCCESS: 'DELETE_ROLE_SUCCESS',
+  DELETE_ROLE_FAILURE: 'DELETE_ROLE_FAILURE',
+
+  EDIT_ROLE_REQUEST: 'EDIT_ROLE_REQUEST',
+  EDIT_ROLE_SUCCESS: 'EDIT_ROLE_SUCCESS',
+  EDIT_ROLE_FAILURE: 'EDIT_ROLE_FAILURE',
+
+  SELECT_ROLE:'SELECT_ROLE',
+
+  EDIT_ROLES_ORDER_REQUEST: 'EDIT_ROLES_ORDER_REQUEST',
+  EDIT_ROLES_ORDER_SUCCESS: 'EDIT_ROLES_ORDER_SUCCESS',
+  EDIT_ROLES_ORDER_FAILURE: 'EDIT_ROLES_ORDER_FAILURE',
+
+  //permissions
+  FETCH_PERMISSIONS_REQUEST: 'FETCH_PERMISSIONS_REQUEST',
+  FETCH_PERMISSIONS_SUCCESS: 'FETCH_PERMISSIONS_SUCCESS',
+  FETCH_PERMISSIONS_FAILURE: 'FETCH_PERMISSIONS_FAILURE',
+
 };
 
 export default types;
