@@ -2,12 +2,9 @@ import React, { Component } from 'react';
 
 import update from 'react/lib/update';
 import map from 'lodash/map';
-import autoBind from 'react-autobind';
 import PermissionItem from './PermissionItem';
-import { DragDropContext } from 'react-dnd';
-import HTML5Backend from 'react-dnd-html5-backend';
 import {SortableContainer, SortableElement, arrayMove} from 'react-sortable-hoc';
-import  './scss/permissions-list.scss';
+import  './scss/xpermissions-list.scss';
 
 const SortablePermissionItem = SortableElement(PermissionItem);
 
@@ -15,12 +12,11 @@ class PermissionsList extends Component {
   render(){
     const { actions, resources, attributes, permissions, order=[] } = this.props;
     return (
-      <div className="permissions-list">
+      <div className="xpermissions-list">
         <h2>permissions:</h2>
         {order.map((e,i)=>
           <SortablePermissionItem
             index={i}
-
             onChange={this.props.onPermissionChange}
             onCreateAction={this.props.onCreateAction}
             onCreateResource={this.props.onCreateResource}
