@@ -1,16 +1,12 @@
-import React, { Component, PropTypes } from 'react';
-import autoBind from 'react-autobind';
+import React, { Component, PropTypes, cloneElement } from 'react';
 import './scss/xtabs.scss';
 import isArray from 'lodash/isArray';
 import Tab from './Tab';
+
 class Tabs extends Component {
-  constructor(props){
-    super(props);
-    autoBind(this);
-    this.state={
-      activeTab:props.initTabIndex,
-      labels:[]
-    };
+  state={
+    activeTab:this.props.initTabIndex,
+    labels:[]
   }
   componentWillMount(){
     const {children}=this.props;
@@ -19,28 +15,28 @@ class Tabs extends Component {
       labels:elements.map(c=>c.props.label)
     });
   }
-  onClick(activeTab){
+  onClick=(activeTab)=>{
     const prevIndex=this.state.activeTab;
     this.setState({activeTab},()=>
-      this.props.onChange({
-        value:
-        prevIndex,
-        curIndex:activeTab
-      })
-    );
+    this.props.onChange({
+      value:
+      prevIndex,
+      curIndex:activeTab
+    }));
   }
-  childrenToArray(children){
+  childrenToArray=(children)=>{
     return isArray(children) ? children : [children];
   }
   render(){
     const {labels, activeTab} = this.state;
-    const {children} = this.props;
+    const {children, style, tabStyle, tabPaneStyle} = this.props;
     const elements = this.childrenToArray(children);
     return (
-      <div className="xtabs">
+      <div className="xtabs" style={style}>
         <ul className="xtabs__bar">
           {labels.map((label,i)=>
             <li
+              style={tabStyle}
               key={i}
               className={'xtabs__tab '+(i==activeTab?'-active':'')}
               onClick={this.onClick.bind(null,i)}>
@@ -48,9 +44,7 @@ class Tabs extends Component {
             </li>
           )}
         </ul>
-        <div className="xtabs__tab-pane">
-          {elements[activeTab]}
-        </div>
+        {cloneElement(elements[activeTab],{style:tabPaneStyle})}
       </div>
     );
   }
@@ -60,9 +54,5 @@ Tabs.defaultProps={
 };
 Tabs.propTypes={
   initTabIndex:PropTypes.number,
-  // children:PropTypes.oneOfType([
-  //   PropTypes.arrayOf(PropTypes.instanceOf(Tab)),
-  //   PropTypes.instanceOf(Tab)
-  // ])
 };
 export default Tabs;
