@@ -3,11 +3,11 @@ import '../static/scss/time-divider.scss';
 
 
 const TimeDivider = ({content})=>{
-    return (
-        <div className="time-divider" >
-            <span>{content}</span>
-        </div>
-    );
+  return (
+    <div className="time-divider" >
+      <span>{content}</span>
+    </div>
+  );
 };
 
 
