@@ -2,18 +2,20 @@ import React, { Component } from 'react';
 import {connect} from 'react-redux';
 import '../static/scss/typing-block.scss';
 
-const TypingIndicator = (props)=>{
+const TypingIndicator = ({
+  users
+})=>{
   return(
-      <div className="typing-block">
-          {props.users.length ? "печатает:" : null}
-          {props.users.map(user=><span key={user.id}> {user.name} </span>)}
-      </div>
+    <div className="typing-block">
+      {users.length ? "печатает:" : null}
+      {users.map(user=><span key={user.id}> {user.name} </span>)}
+    </div>
   );
 };
 
 const mapStateToProps = (state) =>{
-  const channel  = state.pagination.idsByChannel[state.ui.selectedChannel];
-  let {
+  const channel  = state.pagination.idsByChannel[state.ui.params.channelId];
+  const {
     typingUsers=[]
   } = channel||{};
 
