@@ -1,71 +1,52 @@
-import React, { Component } from 'react';
+import React, { Component, PropTypes as PT } from 'react';
 import { connect } from 'react-redux';
 import * as Actions from '../actions/index';
 import SortableChannelList from '../components/ChannelList';
-import {withRouter} from 'react-router';
-import autoBind from 'react-autobind';
 import ModalDialog from '../components/ModalDialog';
 import CreateChannelForm from '../components/CreateChannelForm';
 import ChannelSettings from '../components/ChannelSettings';
-import { arrayMove } from 'react-sortable-hoc';
 import { Scrollbars } from 'react-custom-scrollbars';
-
+import Redirect from 'react-router/Redirect';
+import {getGeneralChannelId} from '../selectors/selectors.js';
 
 class ChannelListContainer extends Component {
-  constructor(props){
-    super(props);
-    autoBind(this);
-    this.state={
-      openCreateChannelModal:false,
-      openChannelSettings:false,
-    };
+  static contextTypes = { router: PT.object.isRequired }
+  state = {
+    openCreateChannelModal:false,
+    openChannelSettings:false,
   }
-  selectChannel(channel){
-    this.props.router.push(`/channels/${this.props.params.serverId}/${channel.id}`);
+  selectChannel=(channel)=>{
+    this.context.router.transitionTo(`/channels/${this.props.serverId}/${channel.id}`);
   }
-  toggleCreateChannelModal(){
+  toggleCreateChannelModal=()=>{
     this.setState({
       openCreateChannelModal:!this.state.openCreateChannelModal
     });
   }
-  submitCreateChannelForm(data){
-    console.log('submitCreateChannelForm');
+  submitCreateChannelForm=(data)=>{
     const {
       description:{value:description},
       name:{value:name}
     } = this.createForm.elements;
-    const {serverId}=this.props.params;
+    const {serverId} = this.props;
     this.props.createChannel({description,name, serverId});
   }
-  onSortEnd({oldIndex, newIndex,...rest}) {
-    const newOrder = arrayMove([...this.props.order], oldIndex, newIndex);
-
-    const piece = oldIndex < newIndex
-    ? newOrder.slice(oldIndex, newIndex+1)
-    : newOrder.slice(newIndex, oldIndex+1);
-
-    let index = oldIndex < newIndex
-    ? oldIndex
-    : newIndex;
-
-    const orderData = piece.map(e=>({id:e, order: ++index}));
-    console.log('orderData',orderData);
-    const {serverId} = this.props.params;
-    this.props.editChannelsOrder(orderData,newOrder,serverId);
+  onSortEnd=({oldIndex, newIndex,...rest})=> {
+    const {serverId} = this.props;
+    this.props.editChannelsOrder(oldIndex, newIndex, serverId);
   }
-  onClickItemSettings(channel){
-    console.log('onClickItemSettings',channel);
+  onClickItemSettings=(channel)=>{
     this.setState({
       openChannelSettings:true,
       ChannelSettingId:channel.id
     });
   }
-  closeChannelSettings(){
+  closeChannelSettings=()=>{
     this.setState({
       openChannelSettings:false
     });
   }
-  handleDeleteChannel(){
+  handleDeleteChannel=()=>{
     this.setState({
       openChannelSettings:false
     },()=>{
@@ -74,14 +55,29 @@ class ChannelListContainer extends Component {
   }
   render(){
     const {
-      channels,selected, params:{channelId},
-      channelsIsFetching, messagesIsFetching , channelsWithNewMessages,
-      order
+      channels, selectedChannelId,
+      order, serverId, generalChannelId, channelsWithNewMessages
     } = this.props;
     const { openCreateChannelModal, openChannelSettings, ChannelSettingId} = this.state;
-    const disableSorting=false;
+    const disableSorting = false;
     return (
       <div style={{width:'100%', height:'100%'}}>
+        <SortableChannelList
+          shouldCancelStart={()=>disableSorting?true:undefined}
+          distance={4}
+          onSortEnd={this.onSortEnd}
+          lockAxis="y"
+          onClickItemSettings={this.onClickItemSettings}
+          onCreateChannel={this.toggleCreateChannelModal}
+          channelsWithNewMessages={channelsWithNewMessages}
+          selected={selectedChannelId}
+          onSelectChannel={this.selectChannel}
+          order={order}
+          channels={channels}/>
+
+        {(channels[selectedChannelId] && channels[selectedChannelId].deleted)
+          && <Redirect to={`/channels/${serverId}/${generalChannelId}`}/>}
+
         {openCreateChannelModal &&
           <ModalDialog
             title={'Создание канала'}
@@ -104,43 +100,32 @@ class ChannelListContainer extends Component {
             style={{content:{background:'#2e3136'}}} >
             <ChannelSettings
               onDeleteChannel={this.handleDeleteChannel}
-              channel={channels[ChannelSettingId]}
-        />
+              channel={channels[ChannelSettingId]}/>
           </ModalDialog>
         }
-        <SortableChannelList
-          shouldCancelStart={()=>disableSorting?true:undefined}
-          distance={4}
-          // pressDelay={50}
-          onSortEnd={this.onSortEnd}
-          lockAxis="y"
-          onClickItemSettings={this.onClickItemSettings}
-          onCreateChannel={this.toggleCreateChannelModal}
-          channelsWithNewMessages={channelsWithNewMessages}
-          selected={channelId}
-          onSelectChannel={this.selectChannel}
-          order={order}
-          channels={channels}/>
 
       </div>
-    );
-  }
-}
+          );
+        }
+      }
 
-const mapStateToProps = (state) =>{
-  console.log(state.ui.selectedServer,'state.ui.selectedServer');
-  // const server  = state.pagination.idsByServer[state.ui.selectedServer];
-  const {channels:order=[]} = state.entities.servers.items[state.ui.selectedServer]||{};
-  // const {
-  //   ids = []
-  // } = server||{};
-  // console.log(server,'server');
-  // console.log(ids,'ids');
-  return {
-    channelsWithNewMessages:state.entities.channels.channelsWithNewMessages,
-    channels:state.entities.channels.items,
-    order
-  };
-};
+      const mapStateToProps = (state) =>{
+        const {
+          channels:order=[]
+        } = state.entities.servers.items[state.ui.params.serverId]||{};
+        const channels = state.entities.channels.items;
+        // const generalChannelId = getGeneralChannelId(channels);
+        // const selectedChannelId = state.ui.params.channelId || generalChannelId ;
+        const selectedChannelId = state.ui.params.channelId ;
+        return {
+          channelSettingsModal:state.ui.modals.channelSettings,
+          serverId:state.ui.params.serverId,
+          // generalChannelId,
+          selectedChannelId,
+          channelsWithNewMessages:state.entities.channels.channelsWithNewMessages,
+          channels,
+          order
+        };
+      };
 
-export default withRouter(connect(mapStateToProps, Actions)(ChannelListContainer));
+      export default connect(mapStateToProps, Actions)(ChannelListContainer);
