@@ -1,23 +1,23 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import * as Actions from '../actions/index';
-import UserList from '../components/UserList';
-
+// import UserList from '../components/UserList';
+ 
 class UserListContainer extends Component {
-
-    render(){
-        const {
-            users, usersOnline
-        } = this.props;
-        return (
-            <UserList users={users} usersOnline={usersOnline}/>
-        );
-    }
+  render(){
+    const {
+      users, usersOnline
+    } = this.props;
+    return (
+      // usersOnline ? <UserList users={users} usersOnline={usersOnline}/> :
+      null
+    );
+  }
 }
 
 const mapStateToProps = (state) =>({
-    users:state.entities.users.items,
-    usersOnline:state.usersOnline.ids
+  users:state.entities.users.items,
+  usersOnline:state.usersOnline.ids
 });
 
 export default connect(mapStateToProps, Actions)(UserListContainer);
