@@ -1,33 +1,32 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router';
 import * as actions from '../../actions/index';
 import ResetPasswordForm from '../../components/auth/ResetPasswordForm';
 
 
 class ResetPasswordPage extends Component {
-    onSubmit(email){
-        console.log('onSubmit ResetPasswordPage ');
-        // this.props.loginUser(creds).then(res=>console.log('1111111111111111111111111',res));
-    }
-    onClickForgotPassword(){
-        // console.log('push to /resetPassword');
-        //
-        // this.props.router.push('/resetPassword');
-    }
-    render(){
-        return (
-            <div>
-                <h1>ResetPasswordPage</h1>
-                {JSON.stringify(this.props.errors)}
-                <ResetPasswordForm  onSubmit={this.onSubmit.bind(this)} />
-            </div>
-        );
-    }
+  onSubmit(email){
+    console.log('onSubmit ResetPasswordPage ');
+    // this.props.loginUser(creds).then(res=>console.log('1111111111111111111111111',res));
+  }
+  onClickForgotPassword(){
+    // console.log('push to /resetPassword');
+    //
+    // this.props.router.push('/resetPassword');
+  }
+  render(){
+    return (
+      <div>
+        <h1>ResetPasswordPage</h1>
+        {JSON.stringify(this.props.errors)}
+        <ResetPasswordForm  onSubmit={this.onSubmit.bind(this)} />
+      </div>
+    );
+  }
 }
 
 const mapStateToProps = (state) =>({
-    errors:state.auth.errors
+  errors:state.auth.errors
 });
 
-export default withRouter(connect(mapStateToProps, actions)(ResetPasswordPage));
+export default connect(mapStateToProps, actions)(ResetPasswordPage);
