@@ -1,4 +1,4 @@
-
+import { AppContainer } from 'react-hot-loader';
 import 'babel-polyfill';
 import 'isomorphic-fetch';
 import {polyfill} from 'es6-promise';
@@ -10,21 +10,34 @@ moment.locale('ru');
 import { render } from 'react-dom';
 import React from 'react';
 import { Provider } from 'react-redux';
-import { Router, Route, browserHistory, IndexRoute } from 'react-router';
 import Store from './store/configureStore';
 import socketEventsListeners from './actions/common/socketEvents';
-
-import routes from './routes';
+import Root from './root';
+// import routes from './routes';
 const store = Store();
 socketEventsListeners(store);
-render(
-  <Provider store={store} >
-    <Router history={browserHistory}>
-      {routes}
-    </Router>
-  </Provider>,
-  document.getElementById('root')
-);
+
+const renderWithHotReload = (RootElement)=> {
+  render(
+    <AppContainer>
+      <Provider store={store}>
+        <RootElement />
+      </Provider>
+    </AppContainer>,
+    document.getElementById('root')
+  );
+};
+
+renderWithHotReload(Root);
+
+if (module.hot) {
+  module.hot.accept('./root', () => {
+    const RootElement = require('./root').default;
+    renderWithHotReload(RootElement);
+  });
+}
+
+
 
 
 
