@@ -8,7 +8,7 @@ export const usersOnline = (state = initialState, action) => {
     switch (action.type) {
 
         case type.UPDATE_USERS_ONLINE:
-        return {ids: action.payload.result};
+        return {ids: action.usersIds};
 
         default: return state;
     }
