@@ -1,29 +1,29 @@
-import React  from 'react';
-
-import { Route, IndexRoute  } from 'react-router';
-import App from './containers/App';
-import Chat from './containers/Chat';
-import LoginPage from './containers/auth/LoginPage';
-
-import ResetPasswordPage from './containers/auth/ResetPasswordPage';
-import ChangePasswordPage from './containers/auth/ChangePasswordPage';
-import SignUpPage from './containers/auth/SignUpPage';
-
-import StartPage from './containers/StartPage';
-import RolesContainer from './components/RolesManager/RolesContainer';
-
-const routes = (
-  <Route  path="/" component={App} >
-    <IndexRoute component={StartPage}/>
-    <Route path="channels/:serverId/:channelId"   component={Chat} />
-    <Route path="channels/:serverId/"   component={Chat} />
-    <Route path="roles" component={RolesContainer} />
-    <Route path="login" component={LoginPage} />
-    <Route path="signup" component={SignUpPage}/>
-    <Route path="resetpassword"  component={ResetPasswordPage}/>
-    <Route path="change_password"  component={ChangePasswordPage}/>
-  </Route>
-);
-
-
-export default routes;
+// import React  from 'react';
+//
+// import { Route, IndexRoute  } from 'react-router';
+// import App from './containers/App';
+// import Chat from './containers/Chat';
+// import LoginPage from './containers/auth/LoginPage';
+//
+// import ResetPasswordPage from './containers/auth/ResetPasswordPage';
+// import ChangePasswordPage from './containers/auth/ChangePasswordPage';
+// import SignUpPage from './containers/auth/SignUpPage';
+//
+// import StartPage from './containers/StartPage';
+// import RolesContainer from './components/RolesManager/RolesContainer';
+//
+// const routes = (
+//   <Route  path="/" component={App} >
+//     <IndexRoute component={StartPage}/>
+//     <Route path="channels/:serverId/:channelId"   component={Chat} />
+//     <Route path="channels/:serverId/"   component={Chat} />
+//     <Route path="roles" component={RolesContainer} />
+//     <Route path="login" component={LoginPage} />
+//     <Route path="signup" component={SignUpPage}/>
+//     <Route path="resetpassword"  component={ResetPasswordPage}/>
+//     <Route path="change_password"  component={ChangePasswordPage}/>
+//   </Route>
+// );
+//
+//
+// export default routes;
