@@ -13,8 +13,8 @@ export default function configureStore(initialState) {
     );
 
     if (module.hot) {
-        module.hot.accept('../reducers', () => {
-            const nextRootReducer = require('../reducers').default;
+        module.hot.accept('../reducers/index', () => {
+            const nextRootReducer = require('../reducers/index').default;
             store.replaceReducer(nextRootReducer);
         });
     }
